+++ src/ui/MemoryMapWindow.cc
@@ -100,8 +100,8 @@ MemoryMapWindow::Draw(kc85& kc) {
     static const ImVec2 page_size(12, 0);
     bool is_kc85_4 = kc.model() == kc85_model::kc85_4;
     bool is_kc85_2 = kc.model() == kc85_model::kc85_2;
-    const int window_height = is_kc85_4 ? 220 : 120;
-    ImGui::SetNextWindowSize(ImVec2(512, window_height), ImGuiSetCond_Always);
+    const float window_height = is_kc85_4 ? 220.0f : 120.0f;
+    ImGui::SetNextWindowSize(ImVec2(512.0f, window_height), ImGuiSetCond_Always);
     if (ImGui::Begin(this->title.AsCStr(), &this->Visible, ImGuiWindowFlags_NoResize|ImGuiWindowFlags_ShowBorders)) {
 
         // draw the background grid
